PhpStan Level 7 : without checkMissingIterableValueType

+++ app/Models/Medecin.php
@@ -80,7 +80,7 @@ class Medecin extends Authenticatable
         return $this->appointments()->where('passed', $bool)->orderBy('date','ASC')->get();
     }
 
-    public function partOfPatient(): int
+    public function partOfPatient() : float
     {
         // recuperer tout les patient du responsable
         $allPatient = 0;
